pyproject.toml
[tool.poetry]
name = "root"
version = "0.1.0"
description = ""
authors = ["elbek <elbekkulmurodov01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
drf-spectacular = "^0.27.2"
djangorestframework-simplejwt = "^5.3.1"
pillow = "^11.0.0"
django-ckeditor-5 = "^0.2.14"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
redis = "^5.2.0"
celery = {extras = ["redis"], version = "^5.4.0"}
django-celery-results = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
